requirements.txt

streamlit
google-cloud-firestore
pandas
